+++ Changes.md
@@ -27,6 +27,7 @@ Breaking Changes
 - Gadget : Added new virtual method and private member variables.
 - ImageGadget : Removed/moved private member variables.
 - OpenColorIOTransform : Made `transform()` method protected rather than public.
+- BackgroundMethodTest : Removed public visibility from private `TestWidget` and `WaitingSlot`.
 
 0.58.3.0 (relative to 0.58.2.0)
 ========